+++ packages/faceted-search/src/components/index.js
@@ -1,4 +1,3 @@
-import { FacetedSearchIcon } from './FacetedSearchIcon';
 import { FacetedSearch } from './FacetedSearch';
 import { AdvancedSearch } from './AdvancedSearch';
 import { BasicSearch } from './BasicSearch';
@@ -13,7 +12,6 @@ import {
 import * as badgeDefinitionTypes from './types/badgeDefinition.type';
 
 export {
-	FacetedSearchIcon,
 	AdvancedSearch,
 	BasicSearch,
 	FacetedSearch,